refactor(rispark): :recycle: Biblioteca refactor
çampos modelo, ajuste endpoints

falta editar archivo por campo

+++ controllers/biblioteca.controller.js
@@ -2,7 +2,8 @@ const { sequelize } = require('../db/connection');
 const { literal } = require('sequelize');
 const { urlFiles } = require('../config/config');
 const BibliotecaModel = require('../models/Biblioteca');
-const { saveResourceMultimedia, deleteResourceMultimedia, deleteFile } = require('../helpers/helpers');
+const { saveResourceMultimedia, deleteResourceMultimedia, deleteFile, generateKeyWord, registerKeyData, validateKeyWord } = require('../helpers/helpers');
+const bcrypt = require('bcrypt')
 
 class BibliotecaController {
   async getFilesLibrary(req, res) {
@@ -16,6 +17,7 @@ class BibliotecaController {
         attributes: [
           'id',
           'estado',
+          'keydata',
           [literal(`(SELECT rm.tipo FROM recursos_multimedia AS rm WHERE rm.id = id_recurso_multimedia)`), 'tipo'],
           [literal(`(SELECT CONCAT('${urlFiles}', rm.recurso) FROM recursos_multimedia AS rm WHERE rm.id = id_recurso_multimedia)`), 'recursoMultimedia'],
         ],
@@ -37,15 +39,23 @@ class BibliotecaController {
   async postFileLibrary(req, res) {
     try {
       return await sequelize.transaction(async (t) => {
-        const { file, token } = req
+        const { file, token, body } = req
+        const { nombre, descripcion, autor } = body
 
         const recursoMultimediaRegistro = await saveResourceMultimedia(file, token?.id);
+        const keydata = await generateKeyWord();
 
         const model = await BibliotecaModel.create({
           idRecursoMultimedia: recursoMultimediaRegistro.id,
           estado: 1,
+          keydata: await bcrypt.hash(keydata, 10),
+          nombre,
+          descripcion,
+          autor,
           createdBy: token.id
         }, { transaction: t })
+
+        await registerKeyData(model.id, recursoMultimediaRegistro.recurso, keydata, 'BI');
 
         return res.status(200).json({ msg: 'success', data: model });
       })
@@ -57,8 +67,11 @@ class BibliotecaController {
   async changeStateFile(req, res) {
     try {
       return await sequelize.transaction(async (t) => {
-        const { token, params } = req;
+        const { token, params, body } = req;
         const file = await BibliotecaModel.findByPk(params.idArchivo);
+
+        const validateKeyData = await validateKeyWord(params.idArchivo, 'BI', body.keydata);
+        if (!validateKeyData) return res.status(400).json({ type: 'error', msg: 'El identificador no concuerda con ningún usuario registrado', status: 400 });
 
         const updateData = {
           estado: file.estado ? 0 : 1,
@@ -77,9 +90,12 @@ class BibliotecaController {
   async deleteFileLibrary(req, res) {
     try {
       return await sequelize.transaction(async (t) => {
-        const { params } = req;
+        const { params, query } = req;
 
         const file = await BibliotecaModel.findByPk(params.idArchivo);
+        const validateKeyData = await validateKeyWord(params.idArchivo, 'BI', query.keydata);
+        if (!validateKeyData) return res.status(400).json({ type: 'error', msg: 'El identificador no concuerda con ningún usuario registrado', status: 400 });
+
         const multimediaFile = await deleteResourceMultimedia(file.idRecursoMultimedia);
 
         deleteFile(multimediaFile, (err) => {
@@ -91,6 +107,14 @@ class BibliotecaController {
       })
     } catch (error) {
       throw error;
+    }
+  }
+
+  async updateFieldsFile(req, res) {
+    try {
+
+    } catch (error) {
+      throw error
     }
   }
 }
